Support SwiftEnumAttribute for closed enum projections
