+++ InteropTests/Tests/SwiftEnumTests.swift
@@ -1,0 +1,2 @@
+// Compile-time test that the type was generated as a Swift enum
+import enum WinRTComponent.SwiftEnum
