+++ InteropTests/WinRTComponent/All.idl
@@ -1,3 +1,5 @@
+// Includes all .idl files so we have a single <Midl Include="All.idl"> entry in the vcxproj.
+// This speeds up the build by avoiding multiple midlrt calls followed by an mdmerge step.
 #include "Arrays.idl"
 #include "ByteBuffers.idl"
 #include "Collections.idl"
@@ -22,4 +24,5 @@
 #include "Strings.idl"
 #include "Structs.idl"
 #include "SwiftAttributes.idl"
+#include "SwiftEnum.idl"
 #include "WeakReferencer.idl"
