+++ InteropTests/WinRTComponent/SwiftEnum.idl
@@ -1,0 +1,12 @@
+namespace WinRTComponent
+{
+    [attributeusage(target_enum)]
+    attribute SwiftEnumAttribute {}
+
+    [SwiftEnum]
+    enum SwiftEnum
+    {
+        Zero = 0,
+        One = 1
+    };
+}
