+++ Support/Sources/WindowsRuntime/ProjectedTypes/TrustLevel.swift
@@ -10,5 +10,5 @@ public struct TrustLevel: Hashable, RawRepresentable, Sendable {
     public static let full = Self(rawValue: 2)
 }
 
-extension TrustLevel: IntegerEnumProjection {
+extension TrustLevel: COM.OpenEnumProjection {
 }
